Changed NSInteger properties of the station object to NSNumber.
Reason for change:

When mapping our dictionaries to our models MTLModel uses NSObject validateValue:forKey:error:
If the properties is of type NSInteger and the value is nil this will return true, it will then
attempt to setValue:forKey with the nil value and cause a crash.

Using NSNumber we protect ourselves against this.